Feature/delete post

+++ lib/resources/firestore_method.dart
@@ -38,17 +38,11 @@ class FirestoreMethod {
         profImage: profImage,
       );
 
-      await _firestore.collection('posts').doc(postId).set({
-        ...post.toJson(),
-        'datePublished':
-            FieldValue.serverTimestamp(), // consistent across clients
-      });
+      _firestore.collection('posts').doc(postId).set(post.toJson());
 
       res = "success";
     } catch (e) {
       res = e.toString();
-      avoidPrint(res);
-      res = "Đã có lỗi xảy ra, vui lòng thử lại";
     }
     return res;
   }
@@ -114,44 +108,27 @@ class FirestoreMethod {
   //Deleting post
   Future<void> deletePost(String postId) async {
     try {
-      await _firestore.collection('posts').doc(postId).delete();
-    } catch (e) {
-      avoidPrint(e.toString());
-    }
-  }
-
-  Future<void> followUser(String uid, String followId) async {
-    try {
-      //fetching all user data
-      DocumentSnapshot snap = await _firestore
-          .collection('users')
-          .doc(uid)
+      // Get the post document to retrieve the postUrl
+      DocumentSnapshot postDoc = await _firestore
+          .collection('posts')
+          .doc(postId)
           .get();
-      //getting following list
-      List following = (snap.data()! as dynamic)['following'];
 
-      //if already following then unfollow
-      if (following.contains(followId)) {
-        //remove follower
-        await _firestore.collection('users').doc(followId).update({
-          'followers': FieldValue.arrayRemove([uid]),
-        });
-        //remove following
-        await _firestore.collection('users').doc(uid).update({
-          'following': FieldValue.arrayRemove([followId]),
-        });
-      } else {
-        //add follower
-        await _firestore.collection('users').doc(followId).update({
-          'followers': FieldValue.arrayUnion([uid]),
-        });
-        //add following
-        await _firestore.collection('users').doc(uid).update({
-          'following': FieldValue.arrayUnion([followId]),
-        });
+      // Check if the document exists
+      if (postDoc.exists) {
+        String postUrl = (postDoc.data() as Map<String, dynamic>)['postUrl'];
+
+        // Delete post collection in Firestore database
+        await _firestore.collection('posts').doc(postId).delete();
+
+        // Delete post's image in storage if it exists
+        if (postUrl.isNotEmpty) {
+          await StorageMethod().deleteImageFromStorage(postUrl);
+        }
       }
     } catch (e) {
       avoidPrint(e.toString());
+      rethrow;
     }
   }
